Type Definition stm32f1::stm32f103::adc1::cr1::DUALMOD_R

type DUALMOD_R = R<u8, DUALMOD_A>;

Reader of field DUALMOD

Implementations

impl DUALMOD_R

pub fn variant(&self) -> Variant<u8, DUALMOD_A>

Get enumerated values variant

pub fn is_independent(&self) -> bool

Checks if the value of the field is INDEPENDENT

pub fn is_regular_injected(&self) -> bool

Checks if the value of the field is REGULARINJECTED

pub fn is_regular_alternate_trigger(&self) -> bool

Checks if the value of the field is REGULARALTERNATETRIGGER

pub fn is_injected_fast_interleaved(&self) -> bool

Checks if the value of the field is INJECTEDFASTINTERLEAVED

pub fn is_injected_slow_interleaved(&self) -> bool

Checks if the value of the field is INJECTEDSLOWINTERLEAVED

pub fn is_injected(&self) -> bool

Checks if the value of the field is INJECTED

pub fn is_regular(&self) -> bool

Checks if the value of the field is REGULAR

pub fn is_fast_interleaved(&self) -> bool

Checks if the value of the field is FASTINTERLEAVED

pub fn is_slow_interleaved(&self) -> bool

Checks if the value of the field is SLOWINTERLEAVED

pub fn is_alternate_trigger(&self) -> bool

Checks if the value of the field is ALTERNATETRIGGER