Type Definition stm32f1::stm32f103::adc1::cr2::JEXTSEL_R

type JEXTSEL_R = R<u8, JEXTSEL_A>;

Reader of field JEXTSEL

Implementations

impl JEXTSEL_R

pub fn variant(&self) -> JEXTSEL_A

Get enumerated values variant

pub fn is_tim1trgo(&self) -> bool

Checks if the value of the field is TIM1TRGO

pub fn is_tim1cc4(&self) -> bool

Checks if the value of the field is TIM1CC4

pub fn is_tim2trgo(&self) -> bool

Checks if the value of the field is TIM2TRGO

pub fn is_tim2cc1(&self) -> bool

Checks if the value of the field is TIM2CC1

pub fn is_tim3cc4(&self) -> bool

Checks if the value of the field is TIM3CC4

pub fn is_tim4trgo(&self) -> bool

Checks if the value of the field is TIM4TRGO

pub fn is_exti15(&self) -> bool

Checks if the value of the field is EXTI15

pub fn is_jswstart(&self) -> bool

Checks if the value of the field is JSWSTART